Moved ExtendedTypeHeader class to the upper level
