+++ msgpack-core/src/main/java/org/msgpack/core/ExtendedTypeHeader.java
@@ -1,0 +1,21 @@
+package org.msgpack.core;
+
+/**
+ * Header of extended type
+ */
+public class ExtendedTypeHeader {
+    private final int type;
+    private final int length;
+    ExtendedTypeHeader(int type, int length) {
+        this.type = type;
+        this.length = length;
+    }
+
+    public int getType() {
+        return type;
+    }
+
+    public int getLength() {
+        return length;
+    }
+}
